refactor(pages/home): extract `<Footer />`
